BUGFIX(oioswift): fix `log_name` directive misplacement since oioswift 1.11.0
 ##### SUMMARY
Since oioswift 1.11.0, the directive `log_name` must be in the `[DEFAULT]`
section of the configuration file `proxy-server.conf`. Before this fix, it was
in another section (`[app:proxy-server]`) and logs were not written to the
oioswift log file in `/var/log`.

As a consequence, statistics gathered from this log file
(through netdata/prometheus/grafana) were inexistant.

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION