CHANGE(oioswift): Activate s3token secret caching
 ##### SUMMARY

To increase performance of the s3 API retrieve and cache s3 secret
from keystone to allow for local validation.

This is a feature from swift >2.20.0 backported to swift3 1.12.20.
open-io/swift3#75

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION

If the new configuration is applied on an old version of s3token, the
parameters will just be ignored.